1610 Lekcija + majas darbs

+++ D6_1610/src/main/java/org/example/Geometrija.java
@@ -1,10 +1,10 @@
 package org.example;
 
 public class Geometrija {
-    public static int triangle (int a, int b, int c) {
+    public static double triangle (double a, double b, double c) {
         return a+b+c;
     }
-    public static int quad(int a, int b, int c, int d) {
+    public static double quad(double a, double b, double c, double d) {
         return a+b+c+d;
     }
 }
